aggiunte le Validation

+++ src/main/java/dev/antoniogrillo/gestioneviaggi/service/impl/MetaServiceImpl.java
@@ -1,81 +1,72 @@
 package dev.antoniogrillo.gestioneviaggi.service.impl;
 
-import dev.antoniogrillo.gestioneviaggi.dto.request.AggiungiMetaDTO;
-import dev.antoniogrillo.gestioneviaggi.dto.response.MetaDTO;
 import dev.antoniogrillo.gestioneviaggi.entity.Meta;
 import dev.antoniogrillo.gestioneviaggi.entity.TipologiaMeta;
-import dev.antoniogrillo.gestioneviaggi.entity.Utente;
-import dev.antoniogrillo.gestioneviaggi.mapper.MetaMapper;
 import dev.antoniogrillo.gestioneviaggi.repository.CriteriaRepository;
 import dev.antoniogrillo.gestioneviaggi.repository.MetaRepository;
 import dev.antoniogrillo.gestioneviaggi.service.def.MetaService;
-import dev.antoniogrillo.gestioneviaggi.service.def.TipologiaMetaService;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class MetaServiceImpl implements MetaService {
 
     private final MetaRepository repository;
-    private final MetaMapper mapper;
     private final CriteriaRepository criteriaRepository;
-    private final TipologiaMetaService tipologiaMetaService;
 
     @Override
-    public List<MetaDTO> getMete(int numeroPagina) {
+    public List<Meta> getMete(int numeroPagina) {
         Sort s= Sort.by("nome").ascending();
         Pageable pageable = PageRequest.of(numeroPagina, 10, s);
-        List<Meta> metas=repository.findAll(pageable).getContent();
-        return mapper.toMetaDTO(metas);
+        return repository.findAll(pageable).getContent();
+
     }
 
     @Override
-    public List<MetaDTO> getMetePerTipologia(long idTipologia, int numeroPagina) {
-        List<Meta> mete=criteriaRepository.findMetePerTipologia(idTipologia,numeroPagina);
-        return mapper.toMetaDTO(mete);
+    public List<Meta> getMetePerTipologia(long idTipologia, int numeroPagina) {
+        return criteriaRepository.findMetePerTipologia(idTipologia,numeroPagina);
     }
 
     @Override
-    public Meta salva(AggiungiMetaDTO request) {
-        Meta m= repository.findByNome(request.nome()).orElse(null);
-        if(m!=null)throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"Nome meta già esistente");
-        List<TipologiaMeta> tipologie=tipologiaMetaService.findAllByIds(request.idTipologie());
-        if(tipologie.isEmpty())throw new ResponseStatusException(HttpStatus.BAD_REQUEST,"Tipologie non valide");
-        m=new Meta();
-        m.setNome(request.nome());
-        m.setTipologie(tipologie);
+    public Meta salva(Meta m) {
         return repository.save(m);
     }
 
     @Override
-    public void elimina(long id) {
-        repository.deleteById(id);
-
+    @Transactional
+    public boolean elimina(long id) {
+        if(repository.existsById(id)) {
+            repository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     @Override
-    public MetaDTO getMeta(long id) {
-        Meta m=repository.findById(id).orElseThrow(()->new ResponseStatusException(HttpStatus.NOT_FOUND,"nessuna meta con questo id"));
-        return mapper.toMetaDTO(m);
+    public Optional<Meta> getMeta(long id) {
+        return repository.findById(id);
     }
 
     @Override
-    public List<MetaDTO> getVisitate(int numeroPagina, Utente utente) {
-        List<Meta> mete=criteriaRepository.findMeteVisitate(utente.getId(),numeroPagina);
-        return mapper.toMetaDTO(mete);
+    public List<Meta> getMetePerIdUtente(int numeroPagina, long idUtente) {
+        return criteriaRepository.findMeteVisitate(idUtente,numeroPagina);
     }
 
     @Override
-    public List<MetaDTO> getDaVisitare(int numeroPagina, Utente utente) {
-        List<Meta> mete=criteriaRepository.findMeteDaVisitare(utente.getId(),numeroPagina);
-        return mapper.toMetaDTO(mete);
+    public List<Meta> getMeteNonVisitatePerIdUtente(int numeroPagina, long idUtente) {
+        return criteriaRepository.findMeteDaVisitare(idUtente,numeroPagina);
+    }
+
+    @Override
+    public Optional<Meta> getByNome(String nome) {
+        return repository.findByNome(nome);
     }
 }
